Explicit upcast

+++ annotate.py
@@ -14,7 +14,7 @@ def annotate_cells(cairo_surface, cell_xyz, radius, rgb, linewidth=5):
 def surface_from_array(a):
     minval = a[a.nonzero()].min()
     maxval = a.max()
-    image_data = a - minval
+    image_data = a.astype(float) - minval
     image_data *= 255.0/(maxval-minval)
     image_data = image_data.astype(np.uint8)
     cairo_image_data = np.dstack([image_data, image_data, image_data, np.ones_like(image_data)*255])
